app/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.0
pyyaml==6.0.1
python-dotenv==1.0.0
geoalchemy2==0.14.1
spacy==3.7.2
yt-dlp==2023.11.16
sentry-sdk==1.38.0
